Baikal: Add config option to discard PCIe initialization
Some PCI devices require the Host Controller to be ready for link
training in a short time after device reset. If the HC appears not
ready this may result in permanent host failure till the next hard
reset.

It seems easier to address the problem in u-boot rather than kernel.
So, if u-boot provides PCIe initialization then initialization in
kernel becomes redundant.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>