clk-baikal: Fix PLL latency
BE_PLL_LATENCY is used as the default delay between checks for
PLL LOCK bit. As a result, minimum latency of frequency switch
is 100000000ns (100ms or 0.1s) which is inacceptable.

According to the Baikal-T datasheet the delay before the first
check for PLL LOCK should be 20us. We define BE_PLL_LATENCY
to that value instead of 100ms.

With this fix measured CPU frequency switch latency appears
to be 150-200us.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>